+++ orinoco-frontend/index.html
@@ -26,7 +26,9 @@
 
     <nav>
       <ul>
-        <li><button id="toOrder">Panier</button></li>
+        <!-- Navigation vers le panier de commande -->
+        <li><a href="/views/order.html">Panier</a></li>
+<!--        <li><button id="toOrder">Panier</button></li> -->
         <li><a href="#">Contact</a></li>
       </ul>
     </nav>
