+++ orinoco-frontend/views/order.html
@@ -26,7 +26,7 @@
 
     <nav>
       <ul>
-        <li><button id="toListOfProducts">Liste des produits</button></li>
+        <li><a href="../index.html">Liste des produits</a></li>
         <li><a href="#">Contact</a></li>
       </ul>
     </nav>
@@ -37,11 +37,15 @@
         
         <table id="orderResult" width='100%' border=true>
           <tr>
+            <th width='10%'>Action</th>
             <th width='10%'>Lien</th>
             <th width='20%'>Référence</th>
             <th width='10%'>Photo</th>
             <th width='10%'>Nom</th>
+            <th width='10%'>Coloris</th>
+            <th width='10%'>Quantité</th>
             <th width='10%'>Prix unitaire</th>
+            <th width='10%'>Sous-total</th>
           </tr>
         </table>
       </article>
