+++ orinoco-frontend/views/thanks.html
@@ -27,7 +27,7 @@
 
     <nav>
       <ul>
-        <li><button id="toListOfProducts">Liste des produits</button></li>
+        <li><a href="../index.html">Liste des produits</a></li>
         <li><a href="#">Contact</a></li>
       </ul>
     </nav>
